UploadVideoViewController is now completely refactored!
